refactor(lab/13): move the life rules/universe out into it's own module, remove cfg_if dependency
per:
https://github.com/rustwasm/rust-webpack-template/commit/dfc704d370cbde0f3ab4f995024136c4a0e4d18a
