+++ src/lab/13-game-of-life/Cargo.toml
@@ -11,11 +11,27 @@ crate-type = ["cdylib"]
 default = ["console_error_panic_hook", "wee_alloc"]
 
 [dependencies]
-cfg-if = "0.1.7"
 console_error_panic_hook = { version = "0.1.6", optional = true }
 wasm-bindgen = "0.2.43"
-web-sys = { version = "0.3.20", features = [ "console" ] }
 wee_alloc = { version = "0.4.4", optional = true }
+
+[dependencies.web-sys]
+version = "0.3.20"
+features = [
+  "CanvasRenderingContext2d",
+  "console",
+  "Document",
+  "Element",
+  "HtmlElement",
+  "HtmlCanvasElement",
+  "Node",
+  "Window",
+]
 
 [dev-dependencies]
 wasm-bindgen-test = "0.2.43"
+
+[profile.release]
+# optimize for size
+# @see: https://doc.rust-lang.org/cargo/reference/manifest.html#the-profile-sections
+opt-level = "s"
