+++ src/lab/13-game-of-life/src/life.rs
@@ -1,0 +1,102 @@
+use std::fmt;
+
+#[repr(u8)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+enum Cell {
+    Dead = 0,
+    Live = 1,
+}
+
+pub struct Universe {
+    width: u32,
+    height: u32,
+    cells: Vec<Cell>,
+}
+
+impl Universe {
+    fn get_index(&self, row: u32, col: u32) -> usize {
+        (row * self.width + col) as usize
+    }
+
+    fn live_neighbors(&self, row: u32, col: u32) -> u8 {
+        let mut n = 0;
+
+        for delta_row in [self.height - 1, 0, 1].iter().cloned() {
+            for delta_col in [self.width - 1, 0, 1].iter().cloned() {
+                if delta_row == 0 && delta_col == 0 {
+                    continue;
+                }
+
+                let neighbor_row = (row + delta_row) % self.height;
+                let neighbor_col = (col + delta_col) % self.width;
+                let i = self.get_index(neighbor_row, neighbor_col);
+
+                n += self.cells[i] as u8;
+            }
+        }
+
+        n
+    }
+
+    pub fn tick(&mut self) {
+        let mut next = self.cells.clone();
+
+        for row in 0..self.height {
+            for col in 0..self.width {
+                let i = self.get_index(row, col);
+                let cell = self.cells[i];
+                let live = self.live_neighbors(row, col);
+
+                let next_cell = match (cell, live) {
+                    (Cell::Live, x) if x < 2 => Cell::Dead,
+                    (Cell::Live, 2) | (Cell::Live, 3) => Cell::Live,
+                    (Cell::Live, x) if x > 3 => Cell::Dead,
+                    (Cell::Dead, 3) => Cell::Live,
+                    (otherwise, _) => otherwise,
+                };
+
+                next[i] = next_cell;
+            }
+        }
+
+        self.cells = next;
+    }
+
+    pub fn new() -> Universe {
+        let width = 24;
+        let height = 24;
+
+        let cells = (0..width * height)
+            .map(|i| {
+                if i % 2 == 0 || i % 7 == 0 {
+                    Cell::Live
+                } else {
+                    Cell::Dead
+                }
+            })
+            .collect();
+
+        Universe {
+            width,
+            height,
+            cells,
+        }
+    }
+
+    pub fn render(&self) -> String {
+        self.to_string()
+    }
+}
+
+impl fmt::Display for Universe {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        for line in self.cells.as_slice().chunks(self.width as usize) {
+            for &cell in line {
+                let symbol = if cell == Cell::Dead { '◻' } else { '◼' };
+                write!(f, "{}", symbol)?;
+            }
+            write!(f, "\n")?;
+        }
+        Ok(())
+    }
+}
